data-pipeline/requirements.txt
# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.16.0
scrapy==2.11.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
polars==0.20.3

# Cleaning & Validation
ftfy==6.1.3
pydantic==2.5.3
jsonschema==4.20.0

# Deduplication
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0
recordlinkage==0.16

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
ipywidgets==8.1.1

# EDA & Statistics
scipy==1.11.4
statsmodels==0.14.1
scikit-learn==1.4.0

# NLP & ML
spacy==3.7.2
transformers==4.36.2
nltk==3.8.1

# Annotation (optionnel - conflits de dépendances, installer séparément si besoin)
# label-studio==1.10.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Utilities
tqdm==4.66.1
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7

# Logging & Monitoring
loguru==0.7.2

# Jupyter
jupyter==1.0.0
jupyterlab==4.0.10
notebook==7.0.6

# Testing
pytest==7.4.3
pytest-cov==4.1.0

